refactor: move image helpers to lib
